src/requirements.txt
django==2.2.3
djangorestframework==3.10
requests==2.22
gunicorn==19.9.0
whitenoise==4.1.3